Moved plotting code
