+++ dna/__main__.py
@@ -11,14 +11,11 @@ import uuid
 import warnings
 import tarfile
 import copy
-import itertools
 
 import numpy as np
 import pandas as pd
-import seaborn as sns
 from tqdm import tqdm
 from tuningdeap import TuningDeap
-import matplotlib.pyplot as plt
 
 from dna import utils
 from dna.data import get_data, preprocess_data, split_data_by_group, group_json_objects
@@ -882,50 +879,6 @@ def aggregate_result_scores(results_to_agg: typing.List[typing.Dict]):
     return agg_scores
 
 
-def create_distribution_plots(agg_results: dict, output_dir: str):
-    scores_by_dataset = agg_results[0]["test_scores"]["scores_by_dataset_id"]
-    if len(list(scores_by_dataset.keys())) == 0:
-        print("No results, skipping")
-        return
-
-    metric_keys = scores_by_dataset[list(scores_by_dataset.keys())[0]].keys()
-    results_dict = {}
-    # aggregates metrics over all datasets
-    for metric_key in metric_keys:
-        new_metric_list = []
-        for index, dataset_name in enumerate(scores_by_dataset.keys()):
-            new_metric_list.append(scores_by_dataset[dataset_name][metric_key])
-        # non-iterable metrics are only non-iterable two layers deep now
-        if isinstance(new_metric_list[0][0], collections.Iterable):
-            new_metric_list = list(zip(*new_metric_list)) # now we have tuples of tuples for each metric
-            flattened_results = [list(sum(tupleOfTuples, ())) for tupleOfTuples in new_metric_list]
-        else:
-            flattened_results = list(itertools.chain(*new_metric_list)) # flatten all the arrays into one array
-        results_dict[metric_key] = flattened_results
-
-    if not os.path.isdir(output_dir):
-        os.makedirs(output_dir)
-
-    # plot aggregate scores in violin plot and save to `output_dir`
-    for metric_key in metric_keys:
-        if isinstance(results_dict[metric_key][0], collections.Iterable):
-            for index in [1, 25, 100, -1]:
-                name_of_index = "all" if index == 1 else str(index)
-                name_of_metric = metric_key.replace("_", " ")
-                ax = sns.violinplot(x=results_dict[metric_key][index])
-                plt.title("Distribution of {} at K={}".format(name_of_metric, name_of_index))
-                plt.xlabel("{} at K={}".format(name_of_metric, index))
-                plt.ylabel("Frequency")
-                plt.savefig(os.path.join(output_dir, "{}-at-{}-violin-plot.png".format(metric_key, name_of_index)))
-                plt.close()
-        else:
-            ax = sns.violinplot(x=results_dict[metric_key])
-            plt.title("Distribution of Metric: {}".format(metric_key))
-            plt.xlabel("{}".format(metric_key))
-            plt.ylabel("Frequency")
-            plt.savefig(os.path.join(output_dir, "{}-violin-plot.png".format(metric_key)))
-            plt.close()
-
 def agg_results_handler(arguments: argparse.Namespace):
     if arguments.results_dir is not None:
         result_paths = get_result_paths_from_dir(arguments.results_dir)
@@ -943,7 +896,7 @@ def agg_results_handler(arguments: argparse.Namespace):
     results_to_agg = [json.load(open(results_path)) for results_path in tqdm(sorted(result_paths))]
 
     agg_scores = aggregate_result_scores(results_to_agg)
-    create_distribution_plots(agg_scores, output_dir)
+    plot.create_distribution_plots(agg_scores, output_dir)
 
     record_run(run_id, git_commit, output_dir, arguments=arguments, model_config=None, scores=agg_scores)
